Update all installed rust binaries with `cargo install <pkgs>` (#17)
* Update all installed rust binaries with `cargo install <pkgs>`

* Vendor deps

We now have one dependency: the `regex` crate. I vendored it.
I also updated the CircleCI config to change the build flag from
`--frozen` to `--locked`. The difference is that `--locked` will make
a network request to determine if `Cargo.lock` is up to date. It will
fail if it's not. The lockfile being out of date is a good error to
catch.